move gcs utils to gcs service
