+++ server/services/storage/google-cloud-storage.service.js
@@ -1,7 +1,12 @@
-const gcsHelpers = require("../../utils/google-cloud-storage");
-const { copyFileToGCS } = gcsHelpers;
+const { Storage } = require("@google-cloud/storage");
+const path = require("path");
 const { upload_path } = require("../../utils/storage");
 const { bucket } = require("../../config");
+
+const storage = new Storage({
+  projectId: process.env.GOOGLE_CLOUD_PROJECT_ID,
+  keyFilename: process.env.GOOGLE_CLOUD_KEYFILE,
+});
 
 exports.saveImage = async function (imageName) {
   const localFilePath = `${upload_path}/${imageName}`;
@@ -13,3 +18,19 @@ exports.getImage = function (name) {
   const url = bucket.name + name;
   return url;
 };
+
+async function copyFileToGCS(localFilePath, bucketName, options) {
+  options = options || {};
+
+  const bucket = storage.bucket(bucketName);
+  const gcsName = path.basename(localFilePath);
+  const file = bucket.file(gcsName);
+
+  await bucket.upload(localFilePath, options);
+  await file.makePublic();
+  return getPublicUrl(bucketName, gcsName);
+}
+
+function getPublicUrl(bucketName, fileName) {
+  `https://storage.googleapis.com/${bucketName}/${fileName}`;
+}
